refactor menu data
