+++ src/data/cateringMenuData/index.js
@@ -1,0 +1,3 @@
+export { plattersCatering } from './plattersCatering';
+export { saladPlattersCatering } from './saladPlattersCatering';
+export { dessertsAndSidesCatering } from './dessertsAndSidesCatering';
