+++ src/data/menuData/index.js
@@ -1,0 +1,6 @@
+export { handHelds } from './handHelds';
+export { largePlates } from './largePlates';
+export { salads } from './salads';
+export { sidesAndSnacks } from './sidesAndSnacks';
+export { kids } from './kids';
+export { dessertsAndBeverages } from './dessertsAndBeverages';
